+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 New functionality:
 * [OLMIS-5138](https://openlmis.atlassian.net/browse/OLMIS-5138): Added `originalRequisition` extended property to Requisition model
 
+Bug fixes:
+* [OLMIS-5847](https://openlmis.atlassian.net/browse/OLMIS-5847): Made that requisition template names are case insensitive unique
+
 7.1.0 / 2018-12-12
 ==================